Rename for ovis_arm and clean up ovis_arm.cpp
