+++ ovis_arm/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(ovis_driver)
+project(ovis_arm)
 
 add_compile_options(-std=c++11)
 find_package(catkin REQUIRED COMPONENTS
@@ -16,7 +16,7 @@ link_directories(${PROJECT_SOURCE_DIR}/lib)
 catkin_package( 
     INCLUDE_DIRS include
     CATKIN_DEPENDS kinova_driver kinova_msgs roscpp std_msgs geometry_msgs ovis_msgs
-    LIBRARIES ovis_driver
+    LIBRARIES ovis_arm
 )
 
 include_directories(
@@ -24,8 +24,8 @@ include_directories(
   ${catkin_INCLUDE_DIRS}
 )
 
-add_executable(${PROJECT_NAME}_node src/ovis_driver.cpp include/ovis_driver.hpp include/kinova/Kinova.API.USBCommandLayerUbuntu.h)
+add_executable(arm src/ovis_arm.cpp include/ovis_arm.hpp include/kinova/Kinova.API.USBCommandLayerUbuntu.h)
 
-add_dependencies(${PROJECT_NAME}_node ${catkin_EXPORTED_TARGETS})
+add_dependencies(arm ${catkin_EXPORTED_TARGETS})
 
-target_link_libraries(${PROJECT_NAME}_node ${catkin_LIBRARIES})
+target_link_libraries(arm ${catkin_LIBRARIES})
