--- ovis_driver/launch/ovis_driver_unity.launch
@@ -1,11 +1,0 @@
-<launch>
-    <arg name="robot_namespace" default="/markhor" />
-
-  <node name="unity_endpoint" pkg="ros_tcp_endpoint" type="default_server_endpoint.py" output="screen">
-      <param name="tcp_ip" type="string" value="192.168.0.107"/>
-      <param name="tcp_port" type="int" value="10000"/>
-  </node>
-
-
-    <node name="ovis_driver_node" pkg="ovis_driver" type="ovis_driver_node" ns="$(arg robot_namespace)" output="screen" />
-</launch>
